secp256k1: Rework k splitting tests.
This reworks the k (scalar) splitting tests to make them more consistent
with modern practices in the code as well as to make the testing
methodology less reliant on the specific implementation details and to
test edge cases.

It also updates the random k splitting tests to ensure new random values
are tested each run as opposed to the existing tests which always test
the same values since they use the same random seed.

Specifically, the following properties are asserted:

- The decomposed scalars must result in the original scalar via the
  equation k = k1 + k2*lambda (mod n)
- Both decomposed scalars must have small magnitude as determined by
  having a maximum bit length of one more than half of the bit size of
  the underlying field