secp256k1/ecdsa: Accept nonce in internal signing.
This refactors the primary logic for the regular ECDSA signing code to a
separate function that accepts an arbitrary nonce.  The primary
motivation is to facilitate upcoming unit tests which aim to fully test
both the positive and negative paths, however, it is also beneficial in
that it provides more flexibility for potentially using other nonce
generation mechanisms in the future should it be necessary.